pyproject.toml
[project]
name = "islamic-digitizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "gradio>=6.12.0",
    "pillow>=12.2.0",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.11.0",
    "torchvision>=0.20.0",
    "torchaudio>=2.5.0",
    "transformers>=5.5.4",
    "peft>=0.19.1",
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "python-multipart>=0.0.26",
]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
